Former-commit-id: 2c01bf452d9a541b0be6cedff8f97151451b0848

+++ Sources/Internal/UI/Styles/UIStyleSheet.cpp
@@ -75,7 +75,7 @@ void UIStyleSheet::RecalculateScore()
     score = 0;
     for (const UIStyleSheetSelector& selector : selectorChain)
     {
-        score += 100000 + selector.classes.size();
+        score += static_cast<uint32>(100000 + selector.classes.size());
         if (selector.name.IsValid())
             score += 100;
         if (!selector.className.empty())
